pyproject.toml
[tool.poetry]
name = "folio_migration_tools"
version = "1.8.3"
description =  "A tool allowing you to migrate data from legacy ILS:s (Library systems) into FOLIO LSP"
authors = ["Theodor Tolstoy <github.teddes@tolstoy.se>", "Lisa Sjögren", "Brooks Travis", "Jeremy Nelson"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/FOLIO-FSE/folio_migration_tools"
repository = "https://github.com/FOLIO-FSE/folio_migration_tools"
keywords = ["FOLIO", "ILS", "LSP", "Library Systems", "MARC21", "Library data"]

[tool.coverage.report]
show_missing = false
fail_under = 10

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["folio_migration_tools"]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 99


[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.9"
folioclient = "^0.50.1"
pyhumps = "^3.7.3"
defusedxml = "^0.7.1"
python-dateutil = "^2.8.2"
folio-uuid = "^0.2.8"
pymarc = "^5.0.0"
pydantic = "^1.10.2"
argparse-prompt = "^0.0.5"
deepdiff = "^6.2.3"
pyaml = "^21.10.1"
httpx = "^0.23.3"
ipykernel = "^6.25.2"
python-i18n = "^0.3.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
lxml = "^4.9.1"
coverage = {extras = ["toml"], version = "^6.5.0"}
pytest-cov = "^4.0.0"
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.982"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.9.23"
flake8-bandit = "^4.1.1"
flake8-isort = "^5.0.0"
flake8-docstrings = "^1.6.0"
darglint = "^1.8.1"
sphinx = "^5.3.0"
sphinx-autodoc-typehints = "^1.19.4"
myst-parser = "^0.18.1"
pandas = "^1.5.3"
types-requests = "^2.28.11.17"
types-python-dateutil = "^2.8.19.11"

[tool.poetry.extras]
docs = ["m2r", "sphinx", "sphinx-autodoc-typehints", "sphinx-rtd-theme", "toml"]
